Fix sign conversion and clang-tidy warnings in core components
SqlStatement.hpp:
- Cast columnIndex to SQLUSMALLINT in BindInputParameter

DataMapper.hpp:
- Change startOffset parameter to SQLUSMALLINT for unsigned column indices

SqlFixedString.hpp:
- Cast indicator to size_t in Resize method

SqlLogger.cpp:
- Simplify stacktrace iteration to avoid implicit conversion

StandardProgressManager.cpp:
- Add NOLINT for IsPinnedTable (intentionally non-static for interface)
- Add deleted copy/move operations to SynchronizedOutputGuard
- Add NOLINT for Update and PrintSummaryLine cognitive complexity
- Use std::cmp_greater/std::cmp_less for signed/unsigned comparisons
- Add explicit size_t casts for vector indexing
- Add parentheses for operator precedence clarity

ddl2cpp.cpp:
- Explicit cast duration to double to avoid precision loss warning

+++ src/Lightweight/DataBinder/SqlFixedString.hpp
@@ -382,7 +382,7 @@ struct SqlBasicStringOperations<SqlFixedString<N, T, Mode>>
 
     static void Resize(ValueType* str, SQLLEN indicator) noexcept
     {
-        str->resize(indicator);
+        str->resize(static_cast<size_t>(indicator));
     }
 
     static void PostProcessOutputColumn(ValueType* result, SQLLEN indicator)
